[backend] move cloning of Projects and Components to Service
